Removed the timout processing program and mvoed it into a background process as part of the main game loop.
Reworked how INotificationHandlers are used - they are by definition in MediatR transient, which means processing logic MUST be stateless.

Implemented a redis-based event producer/consumer that use redis streams as the transport mechanism, allowing a "small" redis server to be used.

Got annoyed with AWS to the point where there is now an EF version of the persistence with the intent of running on postgres.