Lots of changes...
Significant renaming of things to better distinguish between dynamo and postgres implementations.

Implemented a ActiveMQ based event publisher / consumer model as well as using it for delayed processing - this seems to work well (once I get the delay bug sorted out) as it provides the semantics for delivery of kafka we want (consumer groups) with the ability to send a delayed message (so we don't have to rely on SQS).

Implemented a lot of logic to make apps (and apis) work "properly", so they perform clean shutdowns (and disconnects) of services.

Added in a missing font file for the renderer that was used as an embedded resource.