Added additional signaling to ensure that we block applications that are using the Producer or Consumer services from continuing their startup until the services have completely registered themselves with activemq.xml
Changed the ActiveMq model to use a connection/session/[producer|consumer] per thread rather than sharing the connection/session per [producer|consumer] to try and minimise potential race conditions given the threading model.

Changed the topic/queue naming process to just replace non-alpha letters with nothing to make sure we aren't confusing the activemq internal routing patterns with additional characters.

Fixed a mistake in the round end command that was causing exceptions to be thrown when the command is run when there are no guesses in the database.

Added a round extension reason to the roundextended event to indicate the reason why the round was extended which seems important?