Changed implementation of renderer as I hadn't noticed a licence was required, have swapped to rendering SVG's using a FOSS renderer instead.
Tweaks to activemq to make the sessions client acknowledge for the queue consumers to move closer to the fault-tolerance model I want.

Put the minimum answers required back to 1 as it makes  more sense that way - am thinking that round extensions should actually be a factor of the number of answers and how many extensions there have been rather than the fixed values that are currently there.