Made Session, Round and Option aggregates have versions associated, to query these you now need to specify the version number or the handler will return null - this is to detect/prevent race conditions.
Added a distributed locking system into the game event logic to ensure that each session is only being processed by one handler as ActiveMQ does not support a visibility window when reading from a queue, meaning that there are times when (unluckily) the same event is attempted to be processe by two consumers at once. The introduction of versions and locks will mean that the second consumer reading will display a warning as the first consumer has updated the verison number of the session/round.

Updated the use of ActiveMQ to a new base image based on 6.0.1 which is now built from my own Dockerfile and published locally.